edit command for using easily
