+++ .gitignore
@@ -12,7 +12,6 @@ src/__pycache__
 *.pyc
 */convdata/
 *data/
-*.pt
 .vscode
 .wandb
 */__pycache__
