--- command/test.sh
@@ -1,1 +1,0 @@
-python src/main.py --mode test --path . --weight $1
