--- command/train.sh
@@ -1,1 +1,0 @@
-python src/main.py --mode train --path . --epoch $1
