+++ hair_detection/.gitignore
@@ -1,2 +1,1 @@
-model
 __pycache__
